+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## 0.5
 
+### 0.5.9
+
+- Change default config file from `aerich.ini` to `pyproject.toml`. (#197)
+
+  **Upgrade note:**
+    1. Run `aerich init -t config.TORTOISE_ORM`.
+    2. Remove `aerich.ini`.
+
 ### 0.5.8
 
 - Support `indexes` change. (#193)